Refactor : 코드 리펙토링
코드를 리펙토링 하였습니다.
