+++ src/main/java/nbbang/com/nbbang/domain/parties/controller/PartyController.java
@@ -1,4 +1,4 @@
-package nbbang.com.nbbang.domain.party.controller;
+package nbbang.com.nbbang.domain.parties.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -13,14 +13,13 @@ import nbbang.com.nbbang.domain.member.domain.Member;
 import nbbang.com.nbbang.domain.member.service.MemberService;
 import nbbang.com.nbbang.domain.party.domain.Party;
 import nbbang.com.nbbang.domain.party.dto.*;
-import nbbang.com.nbbang.domain.party.dto.many.PartyListResponseDto;
 import nbbang.com.nbbang.domain.party.dto.single.*;
 import nbbang.com.nbbang.domain.party.dto.single.request.PartyChangeGoalNumberRequestDto;
 import nbbang.com.nbbang.domain.party.dto.single.request.PartyRequestDto;
 import nbbang.com.nbbang.domain.party.dto.single.request.PartyStatusChangeRequestDto;
 import nbbang.com.nbbang.domain.party.dto.single.response.PartyIdResponseDto;
 import nbbang.com.nbbang.domain.party.dto.single.response.PartyReadResponseDto;
-import nbbang.com.nbbang.domain.party.service.PartyService;
+import nbbang.com.nbbang.domain.parties.service.PartyService;
 import nbbang.com.nbbang.domain.party.validation.PartyCreateGroup;
 import nbbang.com.nbbang.global.error.exception.CustomIllegalArgumentException;
 import nbbang.com.nbbang.global.interceptor.CurrentMember;
@@ -52,7 +51,7 @@ public class PartyController {
     private final MemberService memberService;
     private final CurrentMember currentMember;
 
-    /*@Operation(summary = "파티 생성", description = "파티를 생성합니다. Place 는 none, sinchon, yeonhui, changcheon 중 하나, 모집 인원은 2~10, 해시태그는 중복 없이 10개 이하")
+    @Operation(summary = "파티 생성", description = "파티를 생성합니다. Place 는 none, sinchon, yeonhui, changcheon 중 하나, 모집 인원은 2~10, 해시태그는 중복 없이 10개 이하")
     @ApiResponse(responseCode = "200", description = "OK",
                     content = @Content(mediaType = "application/json", schema = @Schema(implementation = PartyIdResponseDto.class)))
     @PostMapping
@@ -65,9 +64,9 @@ public class PartyController {
         Party createParty = partyService.create(party, currentMember.id(), partyRequestDto.getHashtags());
         Long partyId = partyService.findIdByParty(createParty);
         return DefaultResponse.res(StatusCode.OK, PartyResponseMessage.PARTY_CREATE_SUCCESS, PartyIdResponseDto.builder().id(partyId).build());
-    }*/
+    }
 
-    @Operation(summary = "파티 생성", description = "파티를 생성합니다. Place 는 none, sinchon, yeonhui, changcheon 중 하나, 모집 인원은 2~10, 해시태그는 중복 없이 10개 이하")
+    /*@Operation(summary = "파티 생성", description = "파티를 생성합니다. Place 는 none, sinchon, yeonhui, changcheon 중 하나, 모집 인원은 2~10, 해시태그는 중복 없이 10개 이하")
     @ApiResponse(responseCode = "200", description = "OK",
             content = @Content(mediaType = "application/json", schema = @Schema(implementation = PartyIdResponseDto.class)))
     @PostMapping
@@ -77,7 +76,7 @@ public class PartyController {
         }
         Long partyId = partyService.createParty(partyRequestDto, currentMember.id());
         return DefaultResponse.res(StatusCode.OK, PartyResponseMessage.PARTY_CREATE_SUCCESS, PartyIdResponseDto.builder().id(partyId).build());
-    }
+    }*/
 
     @Operation(summary = "파티 상세", description = "파티의 상세 정보입니다.")
     @ApiResponse(responseCode = "200", description = "OK",
