+++ src/main/java/nbbang/com/nbbang/domain/parties/controller/PartyResponseMessage.java
@@ -1,6 +1,4 @@
-package nbbang.com.nbbang.domain.party.controller;
-
-import org.webjars.NotFoundException;
+package nbbang.com.nbbang.domain.parties.controller;
 
 public class PartyResponseMessage {
     public static final String PARTY_FIND_SUCCESS = "파티 조회 성공";
